Input validation fix
cin with the extraction operator has space characters as a delimiter by default, which caused problems in my code: Mainly that it caused my input validation system to treat strings seperated with spacebars as different inputs to be checked seperately, resulting in invalid inputs to be validated e.g "32321983123 247"would disqualify the first number stated and qualify the "247", which is not exactly what the user has input. This has been fixed using getline(cin,var) but since it also includes spaces in the input I also needed to create a function to trim leading and trailing space characters.